@@ -2,8 +2,7 @@ package com.nbk.dao.converters;
 
 import com.nbk.dao.domain.account.TransactionTypeEnum;
 
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
+import jakarta.persistence.*;
 import java.util.stream.Stream;
 
 /**
