@@ -2,7 +2,7 @@ package com.nbk.dao.domain.account;
 
 import lombok.Data;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.io.Serializable;
 
 @Entity
