@@ -1,9 +1,9 @@
 package com.nbk.dao.domain.account;
 
+import jakarta.persistence.*;
 import lombok.Data;
 import lombok.Singular;
 
-import javax.persistence.*;
 import java.io.Serializable;
 import java.math.BigDecimal;
 import java.util.ArrayList;
