@@ -1,10 +1,9 @@
 package com.nbk.dto;
 
-import lombok.AllArgsConstructor;
+import jakarta.validation.constraints.NotNull;import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotNull;
 import java.math.BigDecimal;
 
 @Data
